moved some parts
